Rename Layout to MainLayout and put to corresponding folder
